requirements.txt
aiogram==2.25.1
asyncpg==0.27.0
cryptography==41.0.3
python-dotenv==1.0.0
redis==4.5.1
